+++ Podfile
@@ -23,7 +23,16 @@ pod 'Then'
 pod 'ReactorKit'
 
   # Pods for RxSwiftStudyInUIkit
-
+  post_install do |installer|
+    installer.generated_projects.each do |project|
+      project.targets.each do |target|
+          target.build_configurations.each do |config|
+              config.build_settings['IPHONEOS_DEPLOYMENT_TARGET'] = '11.0'
+           end
+      end
+    end
+  end
+  
   target 'RxSwiftStudyInUIkitTests' do
     inherit! :search_paths
     # Pods for testing
